Problem 1: allows user to enroll in a course using a course ID that doessnt exist (fixed
Problem 2: allows for course dropping when course doesnt exist

I think I checked msot of the code with chatgpt and also testing it myself. Do test it yourself and message in the WhatsApp group if you got any improvements. Also kinda formatted the code and make it as formatted as possible